Fix a conversion warning

+++ simdpp/detail/insn/make_const.h
@@ -515,7 +515,7 @@ void construct_eval(V& v, const expr_vec_make_ones& e)
 {
     (void) e;
     expr_vec_make_const<uint64_t,1> e2;
-    e2.a[0] = -1;
+    e2.a[0] = (uint64_t)-1;
     typename V::uint_vector_type u;
     insn::i_make_const(u, e2, 0);
     v = u;
